Changelog > News
You may ask 'why did I do this'

Well, since on the launcher, it's a news thingie, but I don't see it as a news but instead, I see it as a changelog

What's on `Read more >`?
It's a full link to the article. All articles are copy-pasted from the PojavLauncher website repo.

Why is there an excess amount of `<h2></h2>`?
This is to solve something that I don't like.