New thing: pojav_environ
Now, instead of using a lot of env vars, we have a global thing: the pojav_environ!
It provides a global storage between libraries that were loaded multiple times. Currently it is used to store config_renderer, the Pojav window and the main window bundle of gl bridge.
It finally swaps surfaces on my device!